Revert "more fixes after merge"
This reverts commit 5ddaecfbe7e19cc60c9df5af82a4dd5104951488.

+++ packages/marketplace/contracts/RoyaltiesRegistry.sol
@@ -97,7 +97,7 @@ contract RoyaltiesRegistry is IRoyaltiesProvider, OwnableUpgradeable {
     /// @param token address of token
     function clearRoyaltiesType(address token) external {
         _checkOwner(token);
-        royaltiesProviders[token] = uint256(uint160(getProvider(token)));
+        royaltiesProviders[token] = uint(uint160(getProvider(token)));
     }
 
     /// @notice sets royalties for token contract in royaltiesByToken mapping and royalties type = 1
@@ -122,7 +122,7 @@ contract RoyaltiesRegistry is IRoyaltiesProvider, OwnableUpgradeable {
         emit RoyaltiesSetForContract(token, royalties);
     }
 
-    /// @notice returns royalties type from uint256
+    /// @notice returns royalties type from uint
     /// @param data in uint256
     /// @return royalty type
     function _getRoyaltiesType(uint256 data) internal pure returns (uint256) {
@@ -140,7 +140,7 @@ contract RoyaltiesRegistry is IRoyaltiesProvider, OwnableUpgradeable {
     /// @param royaltiesProvider address of royalty provider
     function _setRoyaltiesType(address token, uint256 royaltiesType, address royaltiesProvider) internal {
         require(royaltiesType > 0 && royaltiesType <= ROYALTIES_TYPES_AMOUNT, "wrong royaltiesType");
-        royaltiesProviders[token] = uint256(uint160(royaltiesProvider)) + 2 ** (256 - royaltiesType);
+        royaltiesProviders[token] = uint(uint160(royaltiesProvider)) + 2 ** (256 - royaltiesType);
     }
 
     /// @notice checks if msg.sender is owner of this contract or owner of the token contract
